chapter4_serving_patterns/sync_async_pattern/requirements_proxy.txt
fastapi>=0.60.1
future>=0.18.2
gunicorn>=20.0.4
importlib-metadata>=1.7.0
joblib>=0.15.1
numpy>=1.18.5
Pillow>=7.2.0
psutil>=5.7.0
pydantic>=1.5.1
PyYAML>=5.3.1
redis>=3.5.3
scikit-learn>=0.23.1
starlette>=0.13.4
typing>=3.7.4.1
uvicorn>=0.11.7
uvloop>=0.14.0
httptools>=0.1.1
python-json-logger>=2.0.1
loguru>=0.5.3
requests>=2.25.1
httpx>=0.16.1
tensorflow>=2.4.0
tensorflow-serving-api>=2.4.0